Refactors BoardState to store pieces differently
Now, pieces for red and black are stored as fixed arrays, with Tiles having a weird way of referencing their possession of them.

I think this also fixes a few odd bugs in how the engine works. We'll see!

Basic benchmarking seems to indicate that this improves performance of the engine by 50%. Cool!